+++ src/input_context/input_condition/block_by.rs
@@ -9,18 +9,37 @@ use crate::{
 };
 
 /// Requires another action to not be triggered within the same context.
-///
-/// Could be used for chords to avoid triggering required actions.
 #[derive(Debug)]
 pub struct BlockBy<A: InputAction> {
     /// Action that blocks this condition when active.
     marker: PhantomData<A>,
+
+    /// Wheter to block the state or only the events.
+    ///
+    /// By default set to false.
+    pub events_only: bool,
+}
+
+impl<A: InputAction> BlockBy<A> {
+    /// Block only events.
+    ///
+    /// For details, see [`ConditionKind::Blocker::events_only`].
+    ///
+    /// This can be used for chords to avoid triggering required actions.
+    /// Otherwise, the chord will register the release and cancel itself.
+    pub fn events_only() -> Self {
+        Self {
+            marker: PhantomData,
+            events_only: true,
+        }
+    }
 }
 
 impl<A: InputAction> Default for BlockBy<A> {
     fn default() -> Self {
         Self {
             marker: PhantomData,
+            events_only: false,
         }
     }
 }
@@ -55,7 +74,9 @@ impl<A: InputAction> InputCondition for BlockBy<A> {
     }
 
     fn kind(&self) -> ConditionKind {
-        ConditionKind::Blocker
+        ConditionKind::Blocker {
+            events_only: self.events_only,
+        }
     }
 }